backend/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
psycopg2-binary==2.9.1
pandas==1.3.3
numpy==1.21.2
openpyxl==3.0.9
